add svg graph

+++ src/main/java/createDB.sql
@@ -1,8 +1,9 @@
 drop table if exists result_table cascade;
 
-create table result_table(x int, y double precision, r int, result text);
+create table result_table(x int, y double precision, r double precision, result text);
 
 
 select * from result_table;
+
