+++ src/main/java/utils/HitChecker.java
@@ -16,15 +16,15 @@ public class HitChecker implements Serializable {
 
     }
 
-    private boolean checkTriangle(int xVal, double yVal, int rVal) {
+    private boolean checkTriangle(int xVal, double yVal, double rVal) {
         return (xVal >= 0 && yVal <= 0 && (rVal * rVal) <= (xVal * xVal + yVal * yVal));
     }
 
-    private boolean checkCircle(int xVal, double yVal, int rVal) {
+    private boolean checkCircle(int xVal, double yVal, double rVal) {
         return (xVal >= 0 && yVal >= 0 && (xVal * xVal + yVal * yVal) <= rVal * rVal);
     }
 
-    private boolean checkRectangle(int xVal, double yVal, int rVal) {
+    private boolean checkRectangle(int xVal, double yVal, double rVal) {
         return (xVal <= 0 && yVal >= 0 && rVal <= 0 && xVal <= rVal / 2);
     }
